AU01 Completion
Hello! I have completed AU01. There are 3 sounds which I got from Pixabay, the sounds are rain (that autoplays and loops) a click (that autoplays) and a coin sound effect (that plays when you get the button).
Here is the task -
AU01: Import sound files into the engine. Figure out a way to trigger one-shot sounds, and have looping sounds. You will also need to figure out how to trigger the sounds when the user interacts in the game with either a keypress or mouse click.